+++ AoE2ScenarioRms/errors/__init__.py
@@ -1,2 +1,3 @@
-from .exceptions import LocationNotFoundError, ImproperCreateObjectError, AoE2ScenarioRmsError
+from .exceptions import LocationNotFoundError, ImproperCreateObjectError, AoE2ScenarioRmsError, \
+    InvalidAoE2ScenarioRmsState
 from .warnings import SpawnFailureWarning, AoE2ScenarioRmsError, ImproperCreateObjectWarning